Moved Redux store file
